--- src/main/java/entity/Document.java
@@ -1,69 +1,0 @@
-package entity;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "documents")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Document {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String fileName;
-
-    @Column(nullable = false)
-    private String originalFileName;
-
-    @Column(nullable = false)
-    private String filePath;
-
-    @Column(nullable = false)
-    private Long fileSize;
-
-    @Column(nullable = false)
-    private String mimeType;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private ProcessingStatus status;
-
-    @Lob
-    @Column(columnDefinition = "LONGTEXT")
-    private String extractedText;
-
-    @Lob
-    @Column(columnDefinition = "LONGTEXT")
-    private String interpretedText;
-
-    private String errorMessage;
-
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    private LocalDateTime updatedAt;
-
-    public enum ProcessingStatus {
-        UPLOADED,
-        OCR_PROCESSING,
-        OCR_COMPLETED,
-        AI_PROCESSING,
-        COMPLETED,
-        FAILED
-    }
-}
