+++ src/main/java/law/counsel/CounselApplication.java
@@ -2,8 +2,10 @@ package law.counsel;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
+@EnableAsync
 public class CounselApplication {
 
 	public static void main(String[] args) {
