+++ src/main/java/law/counsel/document/domain/Document.java
@@ -30,10 +30,46 @@ public class Document extends AuditEntity {
     @Column(name = "original_filename")
     private String originalFilename; // 업로드된 원본 파일명
 
+    @Column(name = "file_name")
+    private String fileName; // 저장된 파일명
+
     @ManyToOne
     @JoinColumn(name = "member_id", nullable = false)
     private Member member;
 
     @Enumerated(EnumType.STRING)
     private ContractType contractType; // 계약 유형
+
+    @Column(nullable = false)
+    private String filePath;
+
+    @Column(nullable = false)
+    private Long fileSize;
+
+    @Column(nullable = false)
+    private String mimeType;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private ProcessingStatus status;
+
+    @Lob
+    @Column(columnDefinition = "LONGTEXT")
+    private String extractedText;
+
+    @Lob
+    @Column(columnDefinition = "LONGTEXT")
+    private String interpretedText;
+
+    private String errorMessage;
+
+
+    public enum ProcessingStatus {
+        UPLOADED,
+        OCR_PROCESSING,
+        OCR_COMPLETED,
+        AI_PROCESSING,
+        COMPLETED,
+        FAILED
+    }
 }
